RW-598: Fixed response markup parsing for file sizes greater than Integer.MAX_VALUE.
Modifed the size property on the following classes to support file sizes greater than Integer.MAX_VALUE returned on API response markup:

- RProjectFileDetails
- RProjectResultDetails
- RRepositoryFileDetails

The size property in each case is now of type Long supporting max files sizes of Long.MAX_VALUE.

Relevant unit tests have also been updated to reflect these changes.

+++ src/main/java/com/revo/deployr/client/about/RProjectFileDetails.java
@@ -19,7 +19,7 @@ import java.net.URL;
  */
 public class RProjectFileDetails {
 
-    public RProjectFileDetails(String filename, String descr, String type, int size, URL url) {
+    public RProjectFileDetails(String filename, String descr, String type, long size, URL url) {
         this.filename = filename;
         this.descr = descr;
         this.type = type;
@@ -44,6 +44,6 @@ public class RProjectFileDetails {
     /**
      * Project directory file size.
      */
-    public final int size;
+    public final long size;
 
 }
